+++ core/config/builds/dev/settings.gradle
@@ -13,7 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-include 'shared', 'core', 'plugin:jvm', 'plugin:groovy', 'plugin:javash', 'plugin:ipythonPlugins', 'plugin:node', 'plugin:r', 'plugin:scala', 'plugin:clojure', 'plugin:kdb', 'plugin:cpp', 'plugin:sql','test'
+include 'shared', 'core', 'plugin:jvm', 'plugin:groovy', 'plugin:javash', 'plugin:ipythonPlugins', 'plugin:node', 'plugin:r', 'plugin:scala', 'plugin:clojure', 'plugin:kdb', 'plugin:cpp', 'plugin:sqlsh','test'
 def root = new File(settingsDir, "../../../../")
 project(':shared').projectDir = new File(root, "shared")
 project(':core').projectDir = new File(root, "core")
@@ -28,4 +28,4 @@ project(':plugin:scala').projectDir = new File(root, "/plugin/scala")
 project(':plugin:clojure').projectDir = new File(root, "/plugin/clojure")
 project(':plugin:kdb').projectDir = new File(root, "/plugin/kdb")
 project(':plugin:cpp').projectDir = new File(root, "/plugin/cpp")
-project(':plugin:sql').projectDir = new File(root, "/plugin/sql")
+project(':plugin:sqlsh').projectDir = new File(root, "/plugin/sqlsh")
