+++ plugin/settings.gradle
@@ -43,4 +43,4 @@ project(':plugin:scala').projectDir = file("./scala")
 project(':plugin:clojure').projectDir = file("./clojure")
 project(':plugin:kdb').projectDir = file("./kdb")
 project(':plugin:cpp').projectDir = file("./cpp")
-project(':plugin:cpp').projectDir = file("./sql")
+project(':plugin:sqlsh').projectDir = file("./sqlsh")
