+++ plugin/sqlsh/settings.gradle
